[1.6] Build refactoring

+++ sbin/pack.sh
@@ -1,36 +1,42 @@
 #!/bin/bash
 
-HERE="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
+PROGNAME="${BASH_SOURCE[0]}"
+HERE="$(cd "$(dirname "$PROGNAME")" &>/dev/null && pwd)"
 ROOT=$(cd $HERE/.. && pwd)
 export READIES=$ROOT/deps/readies
 . $READIES/shibumi/defs
 
-cd $ROOT
+SBIN=$ROOT/sbin
 
 export PYTHONWARNINGS=ignore
 
+cd $ROOT
+
 #----------------------------------------------------------------------------------------------
 
-if [[ $1 == --help || $1 == help ]]; then
+if [[ $1 == --help || $1 == help || $HELP == 1 ]]; then
 	cat <<-END
+		Generate RedisTimeSeries distribution packages.
+
 		[ARGVARS...] pack.sh [--help|help]
-		
+
 		Argument variables:
-		MODULE=path       Path to module .so file
+		MODULE=path       Path of module .so
 
 		RAMP=1|0          Build RAMP file
 		DEPS=0|1          Build dependencies file
 		SYM=0|1           Build debug symbols file
 
-		VARIANT=name      Build variant (default: empty)
 		BRANCH=name       Branch name for snapshot packages
-		GITSHA=1          Append Git SHA to shapshot package names
+		VERSION=ver       Version for release packages
+		WITH_GITSHA=1     Append Git SHA to shapshot package names
+		VARIANT=name      Build variant (default: empty)
 
 		ARTDIR=dir        Directory in which packages are created (default: bin/artifacts)
-		BINDIR=dir        Directory in which binaries are found
 
+		JUST_PRINT=1      Only print package names, do not generate
 		VERBOSE=1         Print commands
-		IGNERR=1          Do not abort on error
+		HELP=1            Show help
 
 	END
 	exit 0
@@ -38,54 +44,47 @@ fi
 
 #----------------------------------------------------------------------------------------------
 
-[[ $IGNERR == 1 ]] || set -e
-[[ $V == 1 || $VERBOSE == 1 ]] && set -x
+# RLEC naming conventions
 
-RAMP=${RAMP:-1}
-DEPS=${DEPS:-0}
-SYM=${SYM:-0}
+ARCH=$($READIES/bin/platform --arch)
+[[ $ARCH == x64 ]] && ARCH=x86_64
+
+OS=$($READIES/bin/platform --os)
+[[ $OS == linux ]] && OS=Linux
+
+OSNICK=$($READIES/bin/platform --osnick)
+[[ $OSNICK == trusty ]]  && OSNICK=ubuntu14.04
+[[ $OSNICK == xenial ]]  && OSNICK=ubuntu16.04
+[[ $OSNICK == bionic ]]  && OSNICK=ubuntu18.04
+[[ $OSNICK == focal ]]   && OSNICK=ubuntu20.04
+[[ $OSNICK == jammy ]]   && OSNICK=ubuntu22.04
+[[ $OSNICK == centos7 ]] && OSNICK=rhel7
+[[ $OSNICK == centos8 ]] && OSNICK=rhel8
+[[ $OSNICK == ol8 ]]     && OSNICK=rhel8
+[[ $OSNICK == rocky8 ]]  && OSNICK=rhel8
+
+[[ $OSNICK == bigsur ]]  && OSNICK=catalina
+
+#----------------------------------------------------------------------------------------------
 
 if [[ -z $MODULE || ! -f $MODULE ]]; then
 	eprint "pack: MODULE is not defined or does not refer to a file"
 	exit 1
 fi
 
-[[ -z $BINDIR ]] && BINDIR=$(dirname $MODULE)
-BINDIR=$(cd $BINDIR && pwd)
+RAMP=${RAMP:-1}
+DEPS=${DEPS:-0}
+SYM=${SYM:-0}
 
 [[ -z $ARTDIR ]] && ARTDIR=bin/artifacts
 mkdir -p $ARTDIR $ARTDIR/snapshots
 ARTDIR=$(cd $ARTDIR && pwd)
 
-export ARCH=$($READIES/bin/platform --arch)
-export OS=$($READIES/bin/platform --os)
-export OSNICK=$($READIES/bin/platform --osnick)
-
-# RLEC naming conventions
-[[ $ARCH == x64 ]] && ARCH=x86_64
-[[ $OS == linux ]] && OS=Linux
-if [[ $OSNICK == focal ]]; then
-	OSNICK=ubuntu20.04
-elif [[ $OSNICK == bionic ]]; then
-	OSNICK=ubuntu18.04
-elif [[ $OSNICK == xenial ]]; then
-	OSNICK=ubuntu16.04
-elif [[ $OSNICK == trusty ]]; then
-	OSNICK=ubuntu14.04
-elif [[ $OSNICK == centos7 ]]; then
-	OSNICK=rhel7
-elif [[ $OSNICK == centos8 ]]; then
-	OSNICK=rhel8
-elif [[ $OSNICK == ol8 ]]; then
-	OSNICK=rhel8
-elif [[ $OSNICK == rocky8 ]]; then
-	OSNICK=rhel8	
-fi
-
 export PRODUCT=redistimeseries
+export PRODUCT_LIB=$PRODUCT.so
 export DEPNAMES=""
 
-export PACKAGE_NAME=${PACKAGE_NAME:-${PRODUCT}}
+PACKAGE_NAME=redistimeseries
 
 RAMP_CMD="python3 -m RAMP.ramp"
 
@@ -100,6 +99,8 @@ pack_ramp() {
 	local verspec=${SEMVER}${VARIANT}
 	
 	local fq_package=$stem.${verspec}.zip
+
+	[[ ! -d $ARTDIR ]] && mkdir -p $ARTDIR
 
 	local packfile="$ARTDIR/$fq_package"
 
@@ -117,7 +118,8 @@ pack_ramp() {
 		-e NUMVER -e SEMVER \
 		$ROOT/$rampfile > /tmp/ramp.yml
 	rm -f /tmp/ramp.fname $packfile
-	$RAMP_CMD pack -m /tmp/ramp.yml --packname-file /tmp/ramp.fname --verbose --debug -o $packfile $MODULE >/tmp/ramp.err 2>&1 || true
+	$RAMP_CMD pack -m /tmp/ramp.yml --packname-file /tmp/ramp.fname --verbose --debug \
+		-o $packfile $MODULE >/tmp/ramp.err 2>&1 || true
 	if [[ ! -e $packfile ]]; then
 		eprint "Error generating RAMP file:"
 		>&2 cat /tmp/ramp.err
@@ -130,6 +132,8 @@ pack_ramp() {
 		ln -sf ../$fq_package $snap_package
 	fi
 
+	local packname=`cat /tmp/ramp.fname`
+	echo "Created $packname"
 	cd $ROOT
 }
 
@@ -139,8 +143,8 @@ pack_deps() {
 	local dep="$1"
 
 	local platform="$OS-$OSNICK-$ARCH"
+	local stem=${PACKAGE_NAME}.${dep}.${platform}
 	local verspec=${SEMVER}${VARIANT}
-	local stem=${PACKAGE_NAME}-${dep}.${platform}
 
 	local depdir=$(cat $ARTDIR/$dep.dir)
 
@@ -151,12 +155,12 @@ pack_deps() {
 	{ cd $depdir ;\
 	  cat $ARTDIR/$dep.files | \
 	  xargs tar -c --sort=name --owner=root:0 --group=root:0 --mtime='UTC 1970-01-01' \
-		--transform "s,^,$dep_prefix_dir," 2>> /tmp/pack.err | \
+		--transform "s,^,$dep_prefix_dir," 2> /tmp/pack.err | \
 	  gzip -n - > $tar_path ; E=$?; } || true
 	rm -f $ARTDIR/$dep.prefix $ARTDIR/$dep.files $ARTDIR/$dep.dir
 
 	cd $ROOT
-	if [[ $E != 0 ]]; then
+	if [[ $E != 0 || -s /tmp/pack.err ]]; then
 		eprint "Error creating $tar_path:"
 		cat /tmp/pack.err >&2
 		exit 1
@@ -164,20 +168,21 @@ pack_deps() {
 	sha256sum $tar_path | awk '{print $1}' > $tar_path.sha256
 
 	cd $ARTDIR/snapshots
-	if [[ ! -z $BRANCH ]]; then
+	if [[ -n $BRANCH ]]; then
 		local snap_dep=$stem.${BRANCH}${VARIANT}.tgz
 		ln -sf ../$fq_dep $snap_dep
 		ln -sf ../$fq_dep.sha256 $snap_dep.sha256
 	fi
-	
+
 	cd $ROOT
 }
 
 #----------------------------------------------------------------------------------------------
 
 prepare_symbols_dep() {
+	if [[ ! -f $MODULE.debug ]]; then return 0; fi
 	echo "Preparing debug symbols dependencies ..."
-	echo $BINDIR > $ARTDIR/debug.dir
+	echo $(cd "$(dirname $MODULE)" && pwd) > $ARTDIR/debug.dir
 	echo $MODULE.debug > $ARTDIR/debug.files
 	echo "" > $ARTDIR/debug.prefix
 	pack_deps debug
@@ -186,20 +191,44 @@ prepare_symbols_dep() {
 
 #----------------------------------------------------------------------------------------------
 
-export NUMVER=$(NUMERIC=1 $ROOT/sbin/getver)
-export SEMVER=$($ROOT/sbin/getver)
+NUMVER=$(NUMERIC=1 $SBIN/getver)
+SEMVER=$($SBIN/getver)
 
 if [[ ! -z $VARIANT ]]; then
 	VARIANT=-${VARIANT}
 fi
 
-[[ -z $BRANCH ]] && BRANCH=${CIRCLE_BRANCH:-`git rev-parse --abbrev-ref HEAD`}
+#----------------------------------------------------------------------------------------------
+
+if [[ -z $BRANCH ]]; then
+	BRANCH=$(git rev-parse --abbrev-ref HEAD)
+	# this happens of detached HEAD
+	if [[ $BRANCH == HEAD ]]; then
+		BRANCH="$SEMVER"
+	fi
+fi
 BRANCH=${BRANCH//[^A-Za-z0-9._-]/_}
-if [[ $GITSHA == 1 ]]; then
-	GIT_COMMIT=$(git describe --always --abbrev=7 --dirty="+" 2>/dev/null || git rev-parse --short HEAD)
+if [[ $WITH_GITSHA == 1 ]]; then
+	GIT_COMMIT=$(git rev-parse --short HEAD)
 	BRANCH="${BRANCH}-${GIT_COMMIT}"
 fi
 export BRANCH
+
+#----------------------------------------------------------------------------------------------
+
+if [[ $JUST_PRINT == 1 ]]; then
+	if [[ $RAMP == 1 ]]; then
+		echo "${PACKAGE_NAME}.${OS}-${OSNICK}-${ARCH}.${SEMVER}${VARIANT}.zip"
+	fi
+	if [[ $DEPS == 1 ]]; then
+		for dep in $DEPNAMES; do
+			echo "${PACKAGE_NAME}.${dep}.${OS}-${OSNICK}-${ARCH}.${SEMVER}${VARIANT}.tgz"
+		done
+	fi
+	exit 0
+fi
+
+#----------------------------------------------------------------------------------------------
 
 if [[ $DEPS == 1 ]]; then
 	echo "Building dependencies ..."
@@ -214,7 +243,7 @@ fi
 
 if [[ $RAMP == 1 ]]; then
 	if ! command -v redis-server > /dev/null; then
-		eprint "$0: Cannot find redis-server. Aborting."
+		eprint "Cannot find redis-server. Aborting."
 		exit 1
 	fi
